# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Too Good To Go API",
        "description": "It is an application that allows us to create and manage users and recipes. All users will be able to search for all recipes, they will be able to create and update their own recipes, they will be able to add and delete favorite recipes and they will also be able to see other users' recipes.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://toogoodtogo.test:8080"
        }
    ],
    "paths": {
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "List all available recipes",
                "operationId": "8b9e556d4d747bd60232961951494bfb",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Recipe"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Insert a new recipe",
                "operationId": "46448289940e329a644b891be9c66943",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecipeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Recipe"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseError"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/recipes/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get by id the information for a specific recipe, and also display the information of its ingredients.",
                "operationId": "494913cea16e9bd1ed0d3a25eda1b1f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Recipe"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseError"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Update recipe data",
                "operationId": "61665c36ed1115a4d53992f26035a8fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecipeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Recipe"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseError"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Delete a recipe by id",
                "operationId": "ff2f374ba56017e8d111c4e218868b9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseError"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Recipe": {
                "title": "Recipe",
                "description": "Recipe model",
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "pizza"
                    },
                    "user_id": {
                        "type": "number",
                        "example": "5"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2023-10-23T00:09:16.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2023-10-23T12:33:45.000000Z"
                    }
                },
                "type": "object",
                "example": {
                    "id": 25,
                    "name": "Pizza",
                    "user_id": "6",
                    "created_at": "2023-10-23T00:09:16.000000Z",
                    "updated_at": "2023-10-23T00:10:16.000000Z"
                }
            },
            "RecipeRequest": {
                "title": "RecipeRequest",
                "description": "Recipe Request Model",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "pizza"
                    },
                    "user_id": {
                        "type": "number",
                        "example": "5"
                    }
                },
                "type": "object",
                "example": {
                    "name": "Pizza",
                    "user_id": "6"
                }
            },
...
            "ApiResponse": {
                "title": "ApiResponse",
                "description": "ApiResponse model",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "success"
                    },
                    "status_code": {
                        "type": "number",
                        "example": "200"
                    },
                    "errors": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "data": {
                        "type": "array|object",
                        "example": "[]|{}"
                    }
                },
                "type": "object",
                "example": {
                    "message": "success",
                    "status_code": 200,
                    "errors": false,
                    "data": "[]|{}"
                }
            },
            "ApiResponseError": {
                "title": "ApiResponseError",
                "description": "ApiResponseError model",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "error"
                    },
                    "status_code": {
                        "type": "number",
                        "example": "400"
                    },
                    "errors": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "data": {
                        "type": "array|object",
                        "example": "[]|{}"
                    }
                },
                "type": "object",
                "example": {
                    "message": "error",
                    "status_code": 400,
                    "errors": true,
                    "data": "[]|{}"
                }
            }
        }
    }
}